More testable structure
Triggering a test based on "today is 10 days before the T due date" is kind of hard. By pairing it with "And the daily check for due training runs" then it's clearer what's supposed to happen... There's a date parameter for the daily check which the test code can supply.

Separately, split up a couple of the compound statements to multiple lines so it's clearer how each line maps to a separate regex that calls some code.